+++ .gitmodules
@@ -1,3 +1,4 @@
-[submodule "SFML"]
-	path = SFML
-	url = https://github.com/SFML/SFML
+[submodule "src/SFML"]
+	path = src/SFML
+	url = https://github.com/SFML/SFML.git
+	branch = 2.5.x
