+++ src/main.cpp
@@ -1,6 +1,4 @@
-#include <iostream>
-
-#include "Application.cpp"
+#include "Application.hpp"
 
 int main()
 {
